Make Count of dataset to long type

+++ src/TorchSharp/Dataset.cs
@@ -16,7 +16,7 @@ namespace TorchSharp
                     {
                     }
 
-                    public abstract int Count { get; }
+                    public abstract long Count { get; }
 
                     public abstract Dictionary<string, Tensor> GetTensor(int index);
                 }
